Task 4

+++ internal/api/chat/delete_chat.go
@@ -2,6 +2,7 @@ package chat
 
 import (
 	"context"
+	"fmt"
 	"log"
 
 	desc "github.com/solumD/chat-server/pkg/chat_v1"
@@ -11,6 +12,9 @@ import (
 
 // DeleteChat отправляет запрос на удаление чата в сервисный слой
 func (i *API) DeleteChat(ctx context.Context, req *desc.DeleteChatRequest) (*emptypb.Empty, error) {
+	if req == nil {
+		return nil, fmt.Errorf("req is nil")
+	}
 	_, err := i.chatService.DeleteChat(ctx, req.GetId())
 	if err != nil {
 		log.Println(err)
